requirements.txt
google-api-core>=2.28.1
google-api-python-client>=2.187.0
google-auth>=2.45.0
google-auth-httplib2>=0.3.0
google-cloud-container>=2.61.0
google-cloud-core>=2.5.0
google-cloud-iam>=2.20.0
google-cloud-storage>=3.7.0